requirements.txt
# Python dependencies for Kettler USB to BLE Bridge
# Pure Python implementation with bluez-peripheral

# Web server and WebSocket
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio[client]==5.11.0
python-engineio==4.9.0

# Serial communication
pyserial==3.5

# BLE GATT server (BlueZ D-Bus wrapper)
bluezero==0.8.0

# Event emitter
pyee==11.1.0

# Simple websocket client (for socketio)
simple-websocket==1.0.0

# For Jinja2 templating (Flask dependency)
Jinja2==3.1.3
MarkupSafe==2.1.4

# Werkzeug (Flask dependency)
Werkzeug==3.0.1
